Move sequence stuff to morganey.monad package
(#266)
